Moved makeSingleVersionSet to CrosspareUtils
